Commonize `ViewModelBackStackRecordLocalProvider`, try using retained
