+++ circuit-foundation/build.gradle.kts
@@ -56,6 +56,7 @@ kotlin {
   sourceSets {
     commonMain {
       dependencies {
+        api(libs.androidx.lifecycle.viewModel.compose.jb)
         api(libs.compose.runtime)
         api(libs.compose.foundation)
         api(libs.kotlinx.immutable)
