--- circuit-foundation/src/androidMain/kotlin/com/slack/circuit/foundation/internal/DefaultBackStackRecordLocalProviders.android.kt
@@ -1,10 +1,0 @@
-// Copyright (C) 2025 Slack Technologies, LLC
-// SPDX-License-Identifier: Apache-2.0
-package com.slack.circuit.foundation.internal
-
-import com.slack.circuit.backstack.BackStack
-import com.slack.circuit.backstack.BackStackRecordLocalProvider
-
-internal actual val defaultBackStackRecordLocalProviders:
-  List<BackStackRecordLocalProvider<BackStack.Record>> =
-  listOf(ViewModelBackStackRecordLocalProvider)
